requirements.txt
requests==2.28.1
pandas==2.1.1
python-Levenshtein==0.25.1
